# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Running Trainer App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /week-plans:
    get:
      operationId: listWeekPlans
      tags: [week-plans]
      summary: List week plans
      parameters:
        - name: traineeId
          in: query
          schema:
            type: integer
        - name: weekStart
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of week plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeekPlan"
    post:
      operationId: createWeekPlan
      tags: [week-plans]
      summary: Create a week plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeekPlanInput"
      responses:
        "201":
          description: Created week plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekPlan"
  /week-plans/{id}:
    get:
      operationId: getWeekPlan
      tags: [week-plans]
      summary: Get a week plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Week plan with runs and segments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekPlanDetail"
    put:
      operationId: updateWeekPlan
      tags: [week-plans]
      summary: Update a week plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeekPlanInput"
      responses:
        "200":
          description: Updated week plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekPlan"
    delete:
      operationId: deleteWeekPlan
      tags: [week-plans]
      summary: Delete a week plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /week-plans/{id}/runs:
    post:
      operationId: addRunToWeekPlan
      tags: [week-plans]
      summary: Add a run to a week plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunInput"
      responses:
        "201":
          description: Created run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
  /week-plans/{id}/runs/{runId}:
    put:
      operationId: updateRun
      tags: [week-plans]
      summary: Update a run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: runId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunInput"
      responses:
        "200":
          description: Updated run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
    delete:
      operationId: deleteRun
      tags: [week-plans]
      summary: Delete a run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: runId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────── ANNOUNCEMENTS ────────────────
...
components:
  schemas:
...
    WeekPlan:
      type: object
      properties:
        id:
          type: integer
        traineeId:
          type: integer
        traineeName:
          type: string
        weekStart:
          type: string
          format: date
        notes:
          type: string
        runsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - traineeId
        - weekStart
        - createdAt
    CreateWeekPlanInput:
      type: object
      properties:
        traineeId:
          type: integer
        weekStart:
          type: string
          format: date
        notes:
          type: string
      required:
        - traineeId
        - weekStart
    Run:
      type: object
      properties:
        id:
          type: integer
        weekPlanId:
          type: integer
        name:
          type: string
        runType:
          type: string
        order:
          type: integer
        segments:
          type: array
          items:
            $ref: "#/components/schemas/RunSegment"
      required:
        - id
        - weekPlanId
        - runType
        - order
    CreateRunInput:
      type: object
      properties:
        name:
          type: string
        runType:
          type: string
        order:
          type: integer
        segments:
          type: array
          items:
            $ref: "#/components/schemas/RunSegmentInput"
      required:
        - runType
        - order
    RunSegment:
      type: object
      properties:
        id:
          type: integer
        runId:
          type: integer
        segmentId:
          type: integer
        segmentName:
          type: string
        segmentTemplate:
          type: string
        resolvedText:
          type: string
        order:
          type: integer
      required:
        - id
        - runId
        - order
        - resolvedText
    RunSegmentInput:
      type: object
      properties:
        segmentId:
          type: integer
        resolvedText:
          type: string
        order:
          type: integer
      required:
        - resolvedText
        - order
    WeekPlanDetail:
      type: object
      properties:
        id:
          type: integer
        traineeId:
          type: integer
        traineeName:
          type: string
        weekStart:
          type: string
          format: date
        notes:
          type: string
        runsPerWeek:
          type: integer
        runs:
          type: array
          items:
            $ref: "#/components/schemas/Run"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - traineeId
        - weekStart
        - runs
        - createdAt

    # ─── ANNOUNCEMENTS ───
...