# lib/api-zod/src/generated/types/createRunInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Running Trainer App API
 * OpenAPI spec version: 0.1.0
 */
import type { RunSegmentInput } from "./runSegmentInput";

export interface CreateRunInput {
  name?: string;
  runType: string;
  order: number;
  segments?: RunSegmentInput[];
}