# lib/api-zod/src/generated/types/runSegmentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Running Trainer App API
 * OpenAPI spec version: 0.1.0
 */

export interface RunSegmentInput {
  segmentId?: number;
  resolvedText: string;
  order: number;
}