# lib/api-zod/src/generated/types/weekPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Running Trainer App API
 * OpenAPI spec version: 0.1.0
 */

export interface WeekPlan {
  id: number;
  traineeId: number;
  traineeName?: string;
  weekStart: Date;
  notes?: string;
  runsCount?: number;
  createdAt: Date;
}